Fix flagged deals from being missed
When a deal is flagged after being posted, it is hidden until approved by a moderator. It can be approved at a much later time and due to an old cutoff logic it was being filtered out. Need to change the filter to match the current logic where deals are removed from the DB after two days.